+++ test/copy-non-page-files.test.js
@@ -42,4 +42,22 @@ describe('copyNonPageFiles', () => {
       );
     });
   });
+
+  test('ignores files designated by ignoreWithinPagesDirectory option', () => {
+    copyNonPageFiles({
+      outputDirectory: 'mockOutputDirectory',
+      pagesDirectory: 'mockPagesDirectory',
+      ignoreWithinPagesDirectory: ['**/*.txt', '*.xyz']
+    }).then(() => {
+      expect(cpy).toHaveBeenCalledTimes(1);
+      expect(cpy).toHaveBeenCalledWith(
+        ['**/*.!(js|md|css)', '!**/*.txt', '!*.xyz'],
+        'mockOutputDirectory',
+        {
+          cwd: 'mockPagesDirectory',
+          parents: true
+        }
+      );
+    });
+  });
 });